requirements.txt
requests
selenium
stdiomask
pixivpy
types-requests
types-selenium
